feat: Add message counter UI component
The code changes include adding a new UI component called "message counter". This component is a div element with the data-codicent-type attribute set to "counter" and the data-codicent-search attribute set to "@codicent #log". It automatically fetches the number of messages that contain the search query "@codicent #log" and displays the count.

This commit also includes recent user commits and updates to the getMessages function in codicentjs.js to fetch messages lacking content and include a beforeTimestamp parameter.